requirements.txt
streamlit
numpy
pandas
plotly
Pillow

